requirements.txt
# Core dependencies
openai>=1.3.0
pinecone-client>=2.2.4
python-dotenv>=1.0.0

# PDF processing - primary
pymupdf>=1.23.0
camelot-py>=0.11.0
opencv-python>=4.8.0

# PDF processing - alternatives
unstructured>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# Table handling
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
markdown>=3.4.0

# Web server
Flask>=2.3.0
gunicorn>=21.0.0
Werkzeug>=2.3.0

# Utilities
psutil>=5.9.0
requests>=2.25.0
tqdm>=4.62.0
matplotlib>=3.4.0
networkx>=2.6.0

# Optional dependencies
scikit-learn>=1.2.0

# For development only (commented out for production)
# jupyter>=1.0.0
# pytest>=7.3.1

